test: rename helper and move files
