+++ packages/template-cli/templates/src/ssr/components/ApiPane/index.axe.test.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
 import ApiPane from '.'
-import { accessibilityTestHelper } from '../../axe/accessibilityTestHelper'
+import { accessibilityTestHelper } from '../../__tests__/axe/accessibilityHelper.test'
 
 accessibilityTestHelper(<ApiPane />)
