+++ packages/template-cli/templates/src/ssr/compositions/home/index.axe.test.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
 import Home from '.'
-import { accessibilityTestHelper } from '../../axe/accessibilityTestHelper'
+import { accessibilityTestHelper } from '../../__tests__/axe/accessibilityHelper.test'
 
 accessibilityTestHelper(<Home />)
