+++ packages/template-cli/templates/src/ssr/jest.config.json
@@ -1,5 +1,5 @@
 {
-  "testMatch": ["**/*.test.(ts|tsx)", "!**/*.test.pact.ts"],
+  "testMatch": ["**/*.test.(ts|tsx)", "!**/*.test.pact.ts", "!**/accessibilityHelper.test.tsx"],
   "testPathIgnorePatterns": ["./.next/", "./node_modules/"],
   "transform": {
     "^.+\\.ts$": "ts-jest"
@@ -14,7 +14,7 @@
   },
   "setupFilesAfterEnv": [
     "@testing-library/jest-dom/extend-expect",
-    "./axe/jestAxeSetup.ts"
+    "./__tests__/axe/jestAxeSetup.ts"
   ],
   "globals": {
     "ts-jest": {
